refactor(cli): extract lang parser owner
